simple reflex agent
it is a simple reflex agent that takes input of no. of rooms and which room is dirty or which room is clean , and then agent goes to every room and cleans it and then move to the next room and if the room is already cleaned then it directly moves to the next room and lastly it checks all the rooms again.